Extract Auth struct to own file
